COMP:Fixed old-style cast

+++ VTK/Filtering/vtkCellTypes.cxx
@@ -278,7 +278,7 @@ unsigned long vtkCellTypes::GetActualMemorySize()
     size += this->LocationArray->GetActualMemorySize();
     }
 
-  return (unsigned long) ceil((float)size/1024.0); //kilobytes
+  return static_cast<unsigned long>(ceil(size/1024.0)); //kilobytes
 }
